Debugging iOS build (#566)

+++ include/qinterface.hpp
@@ -205,7 +205,7 @@ public:
 
         if ((rgp == NULL) && (hardware_rand_generator == NULL)) {
             rand_generator = std::make_shared<qrack_rand_gen>();
-            randomSeed = std::time(0);
+            randomSeed = (uint32_t)std::time(0);
             SetRandomSeed(randomSeed);
         } else {
             rand_generator = rgp;
